# Improve scalability of `thread_squad` with tree barrier