--- a/include/sysmakeshift/thread_squad.hpp
+++ b/include/sysmakeshift/thread_squad.hpp
@@ -65,14 +65,13 @@ public:
         //
     class task_context
     {
-        friend detail::thread_squad_thread;
         friend detail::thread_squad_impl;
 
     private:
         detail::thread_squad_impl_base& impl_;
         int threadIdx_;
 
-        explicit task_context(detail::thread_squad_impl_base& _impl, int _threadIdx) noexcept
+        task_context(detail::thread_squad_impl_base& _impl, int _threadIdx) noexcept
             : impl_(_impl), threadIdx_(_threadIdx)
         {
         }
@@ -93,7 +92,7 @@ public:
         gsl_NODISCARD int
         num_threads(void) const noexcept
         {
-            return impl_.numThreads_;
+            return impl_.numThreads;
         }
     };
 
@@ -136,7 +135,7 @@ public:
     gsl_NODISCARD int
     num_threads(void) const
     {
-        return handle_->numThreads_;
+        return handle_->numThreads;
     }
 
         //
@@ -151,7 +150,7 @@ public:
     run(std::function<void(task_context)> action, int concurrency = 0) &
     {
         gsl_Expects(action);
-        gsl_Expects(concurrency >= 0 && concurrency <= handle_->numThreads_);
+        gsl_Expects(concurrency >= 0 && concurrency <= handle_->numThreads);
 
         do_run(std::move(action), concurrency, false);
     }
@@ -168,7 +167,7 @@ public:
     run(std::function<void(task_context)> action, int concurrency = 0) &&
     {
         gsl_Expects(action);
-        gsl_Expects(concurrency >= 0 && concurrency <= handle_->numThreads_);
+        gsl_Expects(concurrency >= 0 && concurrency <= handle_->numThreads);
 
         do_run(std::move(action), concurrency, true);
     }