… test/test-thread_squad.cpp