* Improve scalability of `thread_squad` with tree barrier

  Using a tree-like wait chain significantly reduces contention on large multiprocessor machines.

* Use atomic waits with exponential backoff

* Remove `thread_squad::run_async()`

  This lets us focus on the synchronous use case. It also gets rid of the \<future\> header dependency in \<sysmakeshift/thread_squad.hpp\>.

* Add debugger visualizer for `aligned_buffer<>`