@mbeutel 2020-02-16
Comparing the performance on two different multiprocessor systems for the synthetic benchmark:

- On dual-socket [Intel Xeon Platinum 8168](https://en.wikichip.org/wiki/intel/xeon_platinum/8168) ([Skylake-SP](https://en.wikichip.org/wiki/intel/microarchitectures/skylake_(server)), 48 cores):
  - Before:
    ```
    benchmark name                                  samples       iterations    estimated
                                                    mean          low mean      high mean
                                                    std dev       low std dev   high std dev
    -------------------------------------------------------------------------------
    run                                                       100             1     38.097 ms
                                                       395.725 us    377.964 us     430.63 us
                                                       121.638 us     67.088 us    223.963 us
    ```
  - After:
    ```
    benchmark name                                  samples       iterations    estimated
                                                    mean          low mean      high mean
                                                    std dev       low std dev   high std dev
    -------------------------------------------------------------------------------
    run                                                       100             3     2.6685 ms
                                                        10.035 us      9.152 us     13.642 us
                                                         8.096 us      1.042 us     19.126 us
    ```

- On [Intel Xeon Phi 7210](https://ark.intel.com/content/www/us/en/ark/products/94033/intel-xeon-phi-processor-7210-16gb-1-30-ghz-64-core.html) ([Knights Landing](https://en.wikichip.org/wiki/intel/microarchitectures/knights_landing), single socket, 64 cores, SNC4 Cache mode):
  - Before:
    ```
    benchmark name                                  samples       iterations    estimated
                                                    mean          low mean      high mean
                                                    std dev       low std dev   high std dev
    -------------------------------------------------------------------------------
    run                                                       100             1    553.672 ms
                                                       3.83872 ms    3.35068 ms    4.46247 ms
                                                        2.8124 ms    2.25515 ms    3.76788 ms
    ```
  - After:
    ```
    benchmark name                                  samples       iterations    estimated
                                                    mean          low mean      high mean
                                                    std dev       low std dev   high std dev
    -------------------------------------------------------------------------------
    run                                                       100             5    12.7315 ms
                                                        28.807 us     26.482 us     35.171 us
                                                        17.275 us      1.811 us     36.516 us
    ```

It doesn't mean much because the baseline was terrible, but this is a **39.6×** improvement on Skylake-SP and a **133×** improvement on Knights Landing.